Add optional argument for number of adults in get_price function
PROBLEM: Unable to obtain pricing for more than one adult using pyairbnb
SOLUTION: Add optional argument titled "adults" as int which gets cast as string to ensure continuity with existing sections request (i.e. it asks for "1" instead of just 1). The default value remains 1 if no argument is passed. Argument was appended to end so as not to potentially mess up any existing code people have written that send arguments in specific order (vs. explicitly stating the respective argument)

CAVEAT: I do not know how the pricing will come out if there are extra guest fees. From examining the raw json in chrome as a website visitor, there is a line field for extra guest fee so it is something that I plan to introduce in a following PR once this is merged so I can test. Either ways, merging this PR isn't likely to break anything as all I've added is capability to specify number of guests as optional argument.